Migrated update_entries(), refactored
