+++ .gitignore
@@ -1,3 +1,4 @@
 .env
 env_variables.yaml
 __pycache__
+.vscode
