+++ helper.py
@@ -1,8 +1,10 @@
 import re
 import os
+import json
 from youtube_search import Service
-from dotenv import load_dotenv
-load_dotenv()
+from datetime import datetime
+from config import TMP_FILE
+
 
 
 def extract_channel_id_from_rss_link(str):
@@ -20,3 +22,34 @@ def get_youtube_channel_id_from_custom_name(name):
         10,
         os.getenv('YOUTUBE_DATA_APP_KEY'))
     return service.find_channel_by_custom_url(name)
+
+
+def now():
+    return datetime.utcnow().timestamp()
+
+
+def get_data():
+    if os.path.exists(TMP_FILE):
+        return json.loads(open(TMP_FILE, "r").read())
+    else:
+        return {"last_updated": 0, "currently_crawling": False, "entries": []}
+
+
+def save_data(data):
+    return open(TMP_FILE, "w").write(json.dumps(data))
+
+def is_manim_video(entry):
+    try:
+        check_str = entry["title"] + " " + entry["summary"]
+        return (
+            "manim" in check_str.lower()
+            or "#some" in check_str.lower()
+            or "SoME" in check_str
+        )
+    except:
+        return False
+
+def sanitize(str):
+    re_pattern = re.compile(u'[^\u0000-\uD7FF\uE000-\uFFFF]', re.UNICODE)
+    filtered_string = re_pattern.sub(u'\uFFFD', str)
+    return filtered_string
