+++ pyproject.toml
@@ -17,6 +17,9 @@ flask-cors = "^3.0"
 google-api-python-client = "^2.51"
 gunicorn = "^20.0.4"
 uvicorn = {version = "^0.18.3"}
+Flask-SQLAlchemy = "^2.5.1"
+PyMySQL = "^1.0.2"
+ipdb = "^0.13.9"
 
 
 [build-system]
